refactor: 전략 -> 엔진으로 네이밍 변경
