+++ src/main/kotlin/racingcar/domain/car/engine/implement/DefaultMovingEngine.kt
@@ -1,0 +1,8 @@
+package racingcar.domain.car.engine.implement
+
+import racingcar.domain.car.Position
+import racingcar.domain.car.engine.MovingEngine
+
+class DefaultMovingEngine : MovingEngine {
+    override fun move(position: Position) = position.forward()
+}
