--- src/main/kotlin/racingcar/domain/car/strategy/implement/DefaultMovingStrategy.kt
@@ -1,8 +1,0 @@
-package racingcar.domain.car.strategy.implement
-
-import racingcar.domain.car.Position
-import racingcar.domain.car.strategy.MovingStrategy
-
-class DefaultMovingStrategy : MovingStrategy {
-    override fun move(position: Position) = position.forward()
-}
